# backend-foryou/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ForYou API Documentation",
        "description": "API documentation for the ForYou project",
        "contact": {
            "email": "admin@foryou.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Get list of clientes",
                "description": "Returns list of clientes",
                "operationId": "getClientes",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cliente"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Create a new cliente",
                "description": "Creates a new cliente",
                "operationId": "createCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Get cliente by ID",
                "description": "Returns a single cliente",
                "operationId": "getClienteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cliente to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Update an existing cliente",
                "description": "Updates a cliente",
                "operationId": "updateCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cliente to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Delete a cliente",
                "description": "Deletes a single cliente",
                "operationId": "deleteCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cliente to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Cliente not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "required": [
                    "nombre",
                    "password_hash"
                ],
                "properties": {
                    "id_cliente": {
                        "description": "ID unico del cliente",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nombre": {
                        "description": "Nombre del cliente",
                        "type": "string"
                    },
                    "usuario": {
                        "description": "Nombre de usuario",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Numero de telefono",
                        "type": "string"
                    },
                    "domicilio": {
                        "description": "Domicilio del cliente",
                        "type": "string"
                    },
                    "correo": {
                        "description": "Correo electronico del cliente",
                        "type": "string",
                        "format": "email"
                    },
                    "created_at": {
                        "description": "Fecha de creacion",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualizacion",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ClienteInput": {
                "title": "Cliente Input",
                "required": [
                    "nombre",
                    "password"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del cliente",
                        "type": "string"
                    },
                    "usuario": {
                        "description": "Nombre de usuario (opcional)",
                        "type": "string"
                    },
                    "password": {
                        "description": "Contraseña del cliente (min 8 caracteres)",
                        "type": "string",
                        "format": "password"
                    },
                    "telefono": {
                        "description": "Número de teléfono (opcional)",
                        "type": "string"
                    },
                    "domicilio": {
                        "description": "Domicilio del cliente (opcional)",
                        "type": "string"
                    },
                    "correo": {
                        "description": "Correo electrónico del cliente (opcional, debe ser único)",
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}